test(e2e): add integration tests and fix clippy -D warnings
Task 20: Add pipeline integration tests (3 mock-based e2e tests).
Fix all clippy -D warnings lint errors:
- src/lib.rs: add #![allow(non_snake_case)]
- recovery.rs: remove redundant as u32 cast
- config.rs: vec! -> array, pub fn default() -> impl Default
- dda.rs: #[allow(dead_code)] on output field
- generate_schema.rs: push_str -> push for single char

+++ src/application/recovery.rs
@@ -65,7 +65,7 @@ impl RecoveryStrategy {
     /// Formula: `initial_backoff_ms * 2^consecutive_failures`, capped at
     /// `max_backoff_ms`.  Uses saturating arithmetic to avoid overflow.
     pub fn next_backoff_ms(&self, state: &RecoveryState) -> u64 {
-        let exponent = state.consecutive_failures as u32;
+        let exponent = state.consecutive_failures;
         // 2^exponent, capped — checked_shl returns None on overflow.
         let multiplier: u64 = 1_u64.checked_shl(exponent).unwrap_or(u64::MAX);
         let backoff = self.initial_backoff_ms.saturating_mul(multiplier);
